refactor and fix CS & PHPstan
