+++ src/framework/TestSuite.php
@@ -12,8 +12,8 @@ use RecursiveRegexIterator;
 use ReflectionClass;
 use ReflectionMethod;
 use RegexIterator;
-use ShockedPlot7560\PmmpUnit\framework\attribute\DataProviderAttribute;
-use ShockedPlot7560\PmmpUnit\framework\attribute\DataProviderTest;
+use ShockedPlot7560\PmmpUnit\framework\attribute\dataProvider\DataProviderAttribute;
+use ShockedPlot7560\PmmpUnit\framework\attribute\dataProvider\DataProviderTest;
 use ShockedPlot7560\PmmpUnit\framework\loader\TestSuiteChecker;
 use ShockedPlot7560\PmmpUnit\framework\loader\TestSuiteLoader;
 use ShockedPlot7560\PmmpUnit\utils\Utils;
