+++ tests/pmmpunit/suitetest/normal/tests/attribute/DataProviderAttributeTest.php
@@ -7,10 +7,13 @@ use Generator;
 use Iterator;
 use React\Promise\PromiseInterface;
 use function React\Promise\resolve;
-use ShockedPlot7560\PmmpUnit\framework\attribute\DataProviderAttribute;
+use ShockedPlot7560\PmmpUnit\framework\attribute\dataProvider\DataProviderAttribute;
 use ShockedPlot7560\PmmpUnit\framework\TestCase;
 
 class DataProviderAttributeTest extends TestCase {
+	/**
+	 * @return int[][]
+	 */
 	public function dataProviderArray() : array {
 		return [
 			[1, 2, 3],
@@ -19,6 +22,9 @@ class DataProviderAttributeTest extends TestCase {
 		];
 	}
 
+	/**
+	 * @return PromiseInterface<null>
+	 */
 	#[DataProviderAttribute("dataProviderArray")]
 	public function testDataProviderArray(int $a, int $b, int $expected) : PromiseInterface {
 		$this->assertEquals($expected, $a + $b);
@@ -26,12 +32,18 @@ class DataProviderAttributeTest extends TestCase {
 		return resolve(null);
 	}
 
+	/**
+	 * @return Generator<int[]>
+	 */
 	public function dataProviderGenerator() : Generator {
 		yield [1, 2, 3];
 		yield [2, 3, 5];
 		yield [3, 4, 7];
 	}
 
+	/**
+	 * @return PromiseInterface<null>
+	 */
 	#[DataProviderAttribute("dataProviderGenerator")]
 	public function testDataProviderGenerator(int $a, int $b, int $expected) : PromiseInterface {
 		$this->assertEquals($expected, $a + $b);
@@ -39,6 +51,9 @@ class DataProviderAttributeTest extends TestCase {
 		return resolve(null);
 	}
 
+	/**
+	 * @return Iterator<int[]>
+	 */
 	public function dataProviderIterator() : Iterator {
 		return new ArrayIterator([
 			[1, 2, 3],
@@ -47,6 +62,9 @@ class DataProviderAttributeTest extends TestCase {
 		]);
 	}
 
+	/**
+	 * @return PromiseInterface<null>
+	 */
 	#[DataProviderAttribute("dataProviderIterator")]
 	public function testDataProviderIterator(int $a, int $b, int $expected) : PromiseInterface {
 		$this->assertEquals($expected, $a + $b);
